requirements.txt
sentence-transformers       # embeddings
faiss-cpu                   # vector store
pydantic
mlflow                      # experiment tracking
google-genai
streamlit
uvicorn
